feat: Add dns-timeout configuration
This commit introduces a new configuration option, `dns-timeout`, which allows for setting a specific timeout for DNS requests. This timeout is applied to both UDP and TCP DNS queries.

The implementation identifies DNS queries by checking for traffic on port 53. For UDP, it also verifies that the packet is a query by inspecting the DNS header.

This change provides more granular control over network timeouts, which can be useful in environments with slow DNS servers.
