+++ conf/main.yml
@@ -53,6 +53,8 @@ socks5:
 # connect-timeout: 5000
   # read-write timeout (ms)
 # read-write-timeout: 60000
+  # dns timeout (ms)
+# dns-timeout: 5000
   # stdout, stderr or file-path
 # log-file: stderr
   # debug, info, warn or error
